+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/balance2/BalanceFragment.kt
@@ -1,0 +1,29 @@
+package io.horizontalsystems.bankwallet.modules.balance2
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.compose.ui.platform.ComposeView
+import androidx.compose.ui.platform.ViewCompositionStrategy
+import io.horizontalsystems.bankwallet.core.BaseFragment
+import io.horizontalsystems.bankwallet.modules.balance2.ui.BalanceScreen
+import io.horizontalsystems.core.findNavController
+
+class BalanceFragment : BaseFragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        return ComposeView(requireContext()).apply {
+            setViewCompositionStrategy(
+                ViewCompositionStrategy.DisposeOnLifecycleDestroyed(viewLifecycleOwner)
+            )
+            setContent {
+                BalanceScreen(findNavController())
+            }
+        }
+    }
+}
