+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/main/MainViewPagerAdapter.kt
@@ -2,7 +2,7 @@ package io.horizontalsystems.bankwallet.modules.main
 
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import io.horizontalsystems.bankwallet.modules.balance2.BalanceXxxFragment
+import io.horizontalsystems.bankwallet.modules.balance2.BalanceFragment
 import io.horizontalsystems.bankwallet.modules.market.MarketFragment
 import io.horizontalsystems.bankwallet.modules.settings.main.MainSettingsFragment
 import io.horizontalsystems.bankwallet.modules.transactions.TransactionsFragment
@@ -13,7 +13,7 @@ class MainViewPagerAdapter(fragment: Fragment) : FragmentStateAdapter(fragment.g
 
     override fun createFragment(position: Int) = when (position) {
         0 -> MarketFragment()
-        1 -> BalanceXxxFragment()
+        1 -> BalanceFragment()
         2 -> TransactionsFragment()
         3 -> MainSettingsFragment()
         else -> throw IllegalStateException()
